App/app/requirements.txt
alembic==1.3.0
appnope==0.1.0
asn1crypto==1.3.0
astroid==2.4.1
attrs==19.3.0
Authlib==0.14.3
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.9.0
bleach==3.1.4
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
ecdsa==0.15
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
glob2==0.7
gunicorn==20.0.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.3
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.1
lazy-object-proxy==1.4.3
libarchive-c==2.9
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.5
notebook==6.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
phonenumbers==8.12.3
pickleshare==0.7.5
pkginfo==1.5.0.1
postgres==3.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psutil==5.7.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1
ptyprocess==0.6.0
pyasn1==0.4.8
pycosat==0.6.3
pycparser==2.20
pycryptodome==3.3.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-jose==3.1.0
python-jose-cryptodome==1.3.2
pytz==2020.1
PyYAML==5.3.1
pyzmq==18.1.1
regex==2020.4.4
requests==2.23.0
rsa==4.0
ruamel-yaml==0.15.87
Send2Trash==1.5.0
six==1.13.0
soupsieve==2.0
SQLAlchemy==1.3.10
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.1
zipp==3.1.0
